// src/interfaces/commonInterfaces.ts
export interface Transaction {
  amount: number;
  transactionType: TransactionType;
  paymentType: PaymentType;
  paymentMethod: PaymentMethod;
  paymentId?: string;
  description?: string;
  bookingId?: string;
  createdAt: Date;
  status: PaymentStatus;
}

export interface Post {
  caption: string;
  imageUrl?: string[];
  serviceType: ServiceProvided;
  status?: PostStatus;
  likesCount?: number;
  location?: string;
  createdAt: Date;
  updatedAt: Date;
  vendor_id: mongoose.Types.ObjectId;
  reportCount: number;
  blockReason?: string;
}

// Vendor

export interface Vendor {
  email: string;
  password?: string;
  name: string;
  companyName: string;
  city: string;
  about: string;
  contactinfo: string;
  isActive: boolean;
  isVerified: boolean;
  isAccepted: AcceptanceStatus;
  logo: string;
  imageUrl: string;
  totalBooking: number;
  bookedDates: string[];
  postCount: number;
  refreshToken: string;
  totalRating: number;
  walletBalance: number;
  transactions?: Transaction[];
  posts?: Types.ObjectId[];
  resetPasswordToken?: string;
  resetPasswordExpires?: Date;
  portfolioImages?: string[];
}